Pull out the mock fetched data
