+++ src/mockPayload.js
@@ -1,0 +1,39 @@
+const mockFetchItems = [
+  {
+    title: 'Apple Watermelon',
+    desc: 'Watermelon flavoured apple!',
+    img: 'https://img.icons8.com/dusk/64/000000/unicorn.png'
+  },
+  {
+    title: 'Lemon Water',
+    desc: 'Water flavoured lemon!',
+    img: 'https://img.icons8.com/color/64/000000/fenix.png'
+  },
+  {
+    title: 'Apple Watermelon',
+    desc: 'Watermelon flavoured apple!',
+    img: 'https://img.icons8.com/dusk/64/000000/unicorn.png'
+  },
+  {
+    title: 'Lemon Water',
+    desc: 'Water flavoured lemon!',
+    img: 'https://img.icons8.com/color/64/000000/fenix.png'
+  },
+  {
+    title: 'Apple Watermelon',
+    desc: 'Watermelon flavoured apple!',
+    img: 'https://img.icons8.com/dusk/64/000000/unicorn.png'
+  },
+  {
+    title: 'Lemon Water',
+    desc: 'Water flavoured lemon!',
+    img: 'https://img.icons8.com/color/64/000000/fenix.png'
+  },
+  {
+    title: 'Lemon Water',
+    desc: 'Water flavoured lemon!',
+    img: 'https://img.icons8.com/color/64/000000/fenix.png'
+  },
+ ]
+
+ export default mockFetchItems;
